Only generate keys if they don't exist.

+++ scripts/start.sh
@@ -4,10 +4,10 @@
 if [ -z "$MAIL_HOSTNAME" ]; then
  MAIL_HOSTNAME="mail"
 fi
-if [ -z "$MAIL_HOSTNAME_FQDN"]; then
+if [ -z "$MAIL_HOSTNAME_FQDN" ]; then
   MAIL_HOSTNAME_FQDN="mail.example.com"
 fi
-if [ -z "$POSTMASTER_ADDRESS"]; then
+if [ -z "$POSTMASTER_ADDRESS" ]; then
   POSTMASTER_ADDRESS="postmaster@example.com"
 fi
 
@@ -16,12 +16,14 @@ sed -i -e "s/%MAIL_HOSTNAME_FQDN%/$MAIL_HOSTNAME_FQDN/g" /etc/postfix/main.cf
 sed -i -e "s/%MAIL_HOSTNAME_FQDN%/$MAIL_HOSTNAME_FQDN/g" /etc/opendkim/TrustedHosts
 sed -i -e "s/%POSTMASTER_ADDRESS%/$POSTMASTER_ADDRESS/g" /etc/dovecot/dovecot.conf
 
-openssl genrsa -des3 -passout pass:x -out /etc/ssl/mailcerts/mail.pass.key 2048 && \
-openssl rsa -passin pass:x -in /etc/ssl/mailcerts/mail.pass.key -out /etc/ssl/mailcerts/mail.key
-rm /etc/ssl/mailcerts/mail.pass.key
-openssl req -new -key /etc/ssl/mailcerts/mail.key -out /etc/ssl/mailcerts/mail.csr \
-  -subj "/C=UK/ST=England/L=London/O=OrgName/OU=IT Department/CN=$MAIL_HOSTNAME_FQDN"
-openssl x509 -req -days 365 -in /etc/ssl/mailcerts/mail.csr -signkey /etc/ssl/mailcerts/mail.key -out /etc/ssl/mailcerts/mail_chained.crt
+if [ ! -e /etc/ssl/mailcerts/mail_chained.crt ]; then
+	openssl genrsa -des3 -passout pass:x -out /etc/ssl/mailcerts/mail.pass.key 2048 && \
+	openssl rsa -passin pass:x -in /etc/ssl/mailcerts/mail.pass.key -out /etc/ssl/mailcerts/mail.key
+	rm /etc/ssl/mailcerts/mail.pass.key
+	openssl req -new -key /etc/ssl/mailcerts/mail.key -out /etc/ssl/mailcerts/mail.csr \
+	  -subj "/C=UK/ST=England/L=London/O=OrgName/OU=IT Department/CN=$MAIL_HOSTNAME_FQDN"
+	openssl x509 -req -days 365 -in /etc/ssl/mailcerts/mail.csr -signkey /etc/ssl/mailcerts/mail.key -out /etc/ssl/mailcerts/mail_chained.crt
+fi
 
 # Again set the right permissions (needed when mounting from a volume)
 chown -Rf vmail:vmail /var/vmail/
